maintain

+++ index.js
@@ -57,8 +57,10 @@ client.on('message', message => { // Activities code below when a message is sen
         if (foundInText == false) return; //Checks if the message contains mentions and if it does
     }
 
-    if (foundInText) message.delete();
-
+    if (maintain == 'on') {
+        if (foundInText) message.delete();
+    }
+    
 });
 
 client.login(process.env.token);
